moved proof initialization to reth-optimism-node
